+++ src/components/ui/Modal/ModalButtonGroup.tsx
@@ -31,7 +31,8 @@ export const ModalButtonGroup: FC<ModalButtonGroupProps> = ({
   };
 
   const addCloseBehaviorToButtons = (buttons: JSX.Element[] | JSX.Element) => {
-    if (!buttons || (buttons instanceof Array && buttons.length === 0)) {
+    if (!context.dismissible || !buttons || (buttons instanceof Array && buttons.length === 0)) {
+      context.dismissible && console.warn("the 'dismissible prop prevents buttons from closing the modal");
       return buttons;
     }
     if (!(buttons instanceof Array)) {