+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix icon preventing clicks on `Select` components
 - Fix Github Actions build-storybook error
 - [Docs] Fix ContentShare docs
+- Fix non-overridable Mic prop in `RosterAttendee`
 
 ### Added
 
@@ -27,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a classname to PopOverMenu component for styling access
 - Added forwardRef for Textarea
 - Added `useDevicePermissionStatus` hook as an exported component from the library.
+- Added a "dismissible" prop to Modal to optionally allow persistent modals
 
 ### Changed